refactor: reorganize editor components and logic
