+++ src/features/editor/components/nodes/focus/focus-node.tsx
@@ -8,36 +8,29 @@ import {
 } from "@/components/ui/input-group";
 import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
+import { NodeShell } from "@/features/editor/components/nodes/shared/node-shell";
 import type { FocusNodeType } from "@/features/shared/types/focus-node.types";
-import { cn } from "@/lib/utils";
 
 export function FocusNode({ data, selected }: NodeProps<FocusNodeType>) {
   return (
-    <div
-      className={cn(
-        "h-fit w-84 overflow-hidden rounded-lg border bg-background transition-shadow",
-        selected ? "border-ring shadow-md ring-3 ring-ring/30" : "",
-      )}
-    >
+    <NodeShell selected={selected}>
       <Handle
         type="target"
         position={Position.Top}
         className="border! size-2! border-muted-foreground bg-input!"
       />
-
       <div className="flex items-center gap-2 p-4 py-3">
         <Zap className="size-5 shrink-0 text-muted-foreground" />
         <div className="min-w-0 flex-1 truncate font-medium">
           {data.label === "" ? "Untitled Node" : data.label}
         </div>
         <Badge
-          className="ml-auto bg-pink-100/80 dark:bg-pink-800/25 dark:text-pink-400"
+          className="ml-auto shrink-0 bg-pink-100/80 dark:bg-pink-800/25 dark:text-pink-400"
           variant="secondary"
         >
           Focus
         </Badge>
       </div>
-
       <div className="space-y-6 border-t p-4">
         <div className="space-y-2">
           <Label>Duration</Label>
@@ -60,12 +53,11 @@ export function FocusNode({ data, selected }: NodeProps<FocusNodeType>) {
           </div>
         )}
       </div>
-
       <Handle
         type="source"
         position={Position.Bottom}
         className="border! size-2! border-muted-foreground bg-input!"
       />
-    </div>
+    </NodeShell>
   );
 }
